// frontend/src/app/main/feature/client/infrastructure/models/ClientModel.ts
export interface ClientModel {
    id: string;
    firstName: string;
    lastName: string;
    birthDate: string;
    nit: string;
    businessName: string;
    email: string;
    phoneNumber: string;
    documentType: string;
    documentNumber: string;
    city: string;
    address: string;
    employmentStatus: boolean;
    fundOrigin: string;
    incomeRange: string;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
}